fix test config
