+++ .github/workflows/ci-mac.yml
@@ -86,4 +86,4 @@ jobs:
       run: |
         eval "$(conda shell.bash hook)"
         conda activate conda-env
-        pytest
+        pytest tests
