+++ .github/workflows/ci-windows.yml
@@ -84,4 +84,4 @@ jobs:
       run: |
         eval "$(conda shell.bash hook)"
         conda activate conda-env
-        pytest
+        pytest tests
