+++ src/conftest.py
@@ -4,12 +4,12 @@ Configuration file for doctests (under src/ploomber)
 from pathlib import Path
 import sys
 
-path_to_src = str(Path(__file__, '..', 'src').resolve())
+path_to_src = str(Path(__file__, '..', '..', 'testutils').resolve())
 sys.path.insert(0, path_to_src)
 
-from tests.conftest import fixture_tmp_dir, _path_to_tests  # noqa: E402
+from testutils import fixture_tmp_dir, _path_to_tests  # noqa: E402
 
 
 @fixture_tmp_dir(_path_to_tests() / 'assets' / 'doctests', autouse=True)
-def tmp_nbs():
+def doctests():
     pass
