+++ tests/conftest.py
@@ -3,7 +3,6 @@ Note: tests organized in folders must contain an init file:
 https://github.com/pytest-dev/pytest/issues/3151#issuecomment-360493948
 """
 import subprocess
-import stat
 import base64
 from copy import copy
 import sys
@@ -16,15 +15,20 @@ import os
 import pytest
 from pathlib import Path
 import tempfile
-from functools import wraps
 import test_pkg
 from ploomber.clients import SQLAlchemyClient
 from ploomber import Env
 import pandas as pd
-from glob import iglob
 from ploomber.cli import install, examples
 import posthog
 from unittest.mock import Mock, MagicMock
+
+path_to_src = str(Path(__file__, '..', '..', 'testutils').resolve())
+sys.path.insert(0, path_to_src)
+
+from testutils import (  # noqa: E402
+    fixture_tmp_dir, _path_to_tests, fixture_backup,
+    _fix_all_dot_git_permissions)
 
 
 def git_init():
@@ -67,71 +71,6 @@ def external_access(request, monkeypatch_session):
         monkeypatch_session.setattr(posthog, 'capture',
                                     external_access.get_something)
         yield
-
-
-def _path_to_tests():
-    return Path(__file__).resolve().parent
-
-
-def fixture_tmp_dir(source, **kwargs):
-    """
-    A lot of our fixtures are copying a few files into a temporary location,
-    making that location the current working directory and deleting after
-    the test is done. This decorator allows us to build such fixture
-    """
-
-    # NOTE: I tried not making this a decorator and just do:
-    # some_fixture = factory('some/path')
-    # but didn't work
-    def decorator(function):
-        @wraps(function)
-        def wrapper():
-            old = os.getcwd()
-            tmp_dir = tempfile.mkdtemp()
-            tmp = Path(tmp_dir, 'content')
-            # we have to add extra folder content/, otherwise copytree
-            # complains
-            shutil.copytree(str(source), str(tmp))
-            os.chdir(str(tmp))
-            yield tmp
-
-            # some tests create sample git repos, if we are on windows, we
-            # need to change permissions to be able to delete the files
-            _fix_all_dot_git_permissions(tmp)
-
-            os.chdir(old)
-            shutil.rmtree(tmp_dir)
-
-        return pytest.fixture(wrapper, **kwargs)
-
-    return decorator
-
-
-def fixture_backup(source):
-    """
-    Similar to fixture_tmp_dir but backups the content instead
-    """
-    def decorator(function):
-        @wraps(function)
-        def wrapper():
-            old = os.getcwd()
-            backup = tempfile.mkdtemp()
-            root = _path_to_tests() / 'assets' / source
-            shutil.copytree(str(root), str(Path(backup, source)))
-
-            os.chdir(root)
-
-            yield root
-
-            os.chdir(old)
-
-            shutil.rmtree(str(root))
-            shutil.copytree(str(Path(backup, source)), str(root))
-            shutil.rmtree(backup)
-
-        return pytest.fixture(wrapper)
-
-    return decorator
 
 
 @pytest.fixture(scope='session')
@@ -186,20 +125,6 @@ def backup_simple():
 @fixture_backup('online')
 def backup_online():
     pass
-
-
-def _fix_dot_git_permissions(path):
-    for root, dirs, files in os.walk(path):
-        for dir_ in dirs:
-            os.chmod(Path(root, dir_), stat.S_IRWXU)
-        for file_ in files:
-            os.chmod(Path(root, file_), stat.S_IRWXU)
-
-
-def _fix_all_dot_git_permissions(tmp):
-    if os.name == 'nt':
-        for path in iglob(f'{tmp}/**/.git', recursive=True):
-            _fix_dot_git_permissions(path)
 
 
 @pytest.fixture()
